Moving configuration base out of axe.windows
