+++ src/AccessibilityInsights.SharedUx/Settings/AppLayout.cs
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Core.Bases;
 using System.Windows;
 
 namespace AccessibilityInsights.SharedUx.Settings
