+++ src/AccessibilityInsights.SharedUx/Settings/ConfigurationManager.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Desktop.Settings;
 using Axe.Windows.Desktop.UIAutomation;
-using Axe.Windows.RuleSelection;
 using AccessibilityInsights.SetupLibrary;
 using System;
 using System.IO;
