+++ src/AccessibilityInsights.SharedUx/Settings/RecorderSetting.cs
@@ -1,10 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Core.Bases;
-using Axe.Windows.Core.Types;
-using Axe.Windows.Desktop.Types;
 using System.Collections.Generic;
-using System.Linq;
 using UIAutomationClient;
 
 namespace AccessibilityInsights.SharedUx.Settings
